+++ packages/design-tokens/src/utils/utils.ts
@@ -16,3 +16,19 @@ export const isNonEmpty = (s: unknown) => {
 
 export const isFulfilled = <T>(result: PromiseSettledResult<T>): result is PromiseFulfilledResult<T> =>
     result.status === 'fulfilled'
+
+export const uniq = <T>(array: T[]): T[] => {
+    const seen = new Set<T>()
+
+    return array.filter((item) => {
+        // `item`이 string 타입일 경우 toLowerCase() 적용
+        const normalizedItem = (typeof item === 'string' ? item.toLowerCase() : item) as T
+
+        if (seen.has(normalizedItem)) {
+            return false
+        }
+
+        seen.add(normalizedItem)
+        return true
+    })
+}
